init commit

+++ routes/api.php
@@ -16,12 +16,15 @@ use App\Http\Controllers\Api\Category_listcontroller;
 |
 */
 
-Route::get('/user',[User_listController::class , 'index']);
+Route::post('/user',[User_listController::class , 'index']);
+Route::post( '/signup',[User_listController::class , 'store']);
 
-Route::get('/products',[Product_listController::class , 'index']);
-Route::get('/categories',[Category_listcontroller::class , 'index']);
-Route::post('/addproducts',[Product_listController::class, 'store']);
-Route::delete('delete-product/{id}' , [Product_listController::class, 'delete']);
+Route::get('/products',[Product_listController::class , 'index'])->middleware('auth:api');
+Route::get('/categories',[Category_listcontroller::class , 'index'])->middleware('auth:api');
+Route::post('/addproducts',[Product_listController::class, 'store'])->middleware('auth:api');
+Route::delete('delete-product/{id}' , [Product_listController::class, 'delete'])->middleware('auth:api');
+Route::get('/viewproduct/{id}',[Product_listController::class , 'viewProduct'])->middleware('auth:api');
+Route::get('/viewproduct/{cat_id}',[Product_listController::class,'ShowDataByCategory'])->middleware('auth:api');
 
 //Route::middleware('auth:api')->get('/user', function (Request $request) {
 //    return $request->user();
